+++ components/board/BestArticle.tsx
@@ -3,9 +3,11 @@ import styles from "@/styles/BestArticle.module.scss";
 import Image from "next/image";
 import Link from "next/link";
 import { getDate } from "@/utils/getDate";
+import useDeviceSize from "@/hooks/useDeviceSize";
+import { useEffect, useState } from "react";
+import axios from "@/lib/axios";
 
 // NOTE - 베스트 게시글
-
 interface BestArticleProps {
   article: Article;
 }
@@ -57,13 +59,34 @@ function BestArticle({ article }: BestArticleProps) {
 }
 
 // NOTE - 베스트 게시글 리스트
-interface BestArticleListProps {
-  bestArticleList: Article[];
-}
 
-export default function BestArticleList({
-  bestArticleList,
-}: BestArticleListProps) {
+export default function BestArticleList() {
+  const { isDesktop, isMobile, isTablet, bestPageSizeCount } = useDeviceSize();
+  const [bestArticleList, setBestArticleList] = useState<Article[]>([]);
+  const [bestPageSize, setBestPageSize] = useState(bestPageSizeCount);
+
+  async function getBestArticleList() {
+    const res = await axios.get(
+      `/articles?pageSize=${bestPageSize}&orderBy=like`
+    );
+    const nextArticleList = res.data.list;
+    setBestArticleList(nextArticleList);
+  }
+
+  useEffect(() => {
+    getBestArticleList();
+  }, [bestPageSize]);
+
+  useEffect(() => {
+    if (isMobile) {
+      setBestPageSize(1);
+    } else if (isTablet) {
+      setBestPageSize(2);
+    } else if (isDesktop) {
+      setBestPageSize(3);
+    }
+  }, [isMobile, isTablet, isDesktop]);
+
   return (
     <div className={styles["best-article-list-container"]}>
       {bestArticleList.map((article) => {
